creates ip component
