--- src/app/ip/ip.tsx
@@ -1,85 +1,0 @@
-
-
-export default function IP() {
-  return (
-    <div className='flex flex-col items-center justify-center w-full md:max-w-[720px] max-w-[320px] h-full mx-auto flex-1 px-4 text-black pt-48 pb-32'>
-      <p className='sm:text-7xl lg:text-8xl text-4xl uppercase'>IP Rights</p>
-      <p>Last Updated: April 28, 2024</p>
-      <p className='self-start pt-8'>Background</p>
-      <p className='font-vista-sans'>
-        The subject of Intellectual Property (IP) rights has been an interesting
-        and evolving topic in the Non-fungible Token (NFT) space. Froggy Friends
-        has reviewed numerous currently utilized intellectual property documents
-        and standards in the respective copyright and intellectual fields and will give
-        credit to and cite industry leaders where applicable. Pudgy Penguins and
-        World of Women (WOW) are by far the current industry leader in IP
-        rights. Pudgy Penguins governing law body currently is documented under
-        the State of Delaware, while WOW is governed in accordance with the laws
-        of France. Froggy Friends IP is currently primarily governed under the State
-        of California law. For the foreseeable future, we will try to adhere to the
-        gold standard set forth by Pudgy Penguins and WOW on IP right. This is
-        important so that Froggy Friends can evolve with the rapidly changing new
-        technology as legal precedence and definitions potentially change current
-        understanding of intellectual property rights in accordance with U.S. law.
-        At the core, even though NFT technology is relatively new, the IP laws
-        governing this new field are not. Basic definitions, along with cited
-        materials and legal documents will be given below to help the reader
-        understand certain common legal terms.
-      </p>
-      <p className='self-start pt-8'>1. DEFINITIONS</p>
-      <p className='font-vista-sans py-2'>
-        <strong className='font-hwt-artz'>“INTELLECTUAL PROPERTY (IP) LAW” </strong>IS A SET OF
-        EXCLUSIVE RIGHTS TO EXCLUDE OTHERS FROM MAKING,
-        COPYING, OR USING CERTAIN INTANGIBLE CREATIONS OF THE
-        HUMAN MIND. THE U.S. CONSTITUTION PROVIDES CONGRESS
-        WITH TWO POWERS RELEVANT TO IP RIGHTS. FIRST, THE IP
-        CLAUSE EMPOWERS CONGRESS TO GRANT “AUTHORS AND
-        INVENTORS THE EXCLUSIVE RIGHT TO THEIR RESPECTIVE
-        WRITINGS AND DISCOVERIES” IN ORDER TO “PROMOTE THE
-        PROGRESS OF SCIENCE AND USEFUL ARTS.” SECOND, THE
-        COMMERCE CLAUSE ALLOWS CONGRESS TO REGULATE
-        INTERSTATE AND FOREIGN COMMERCE. GENERALLY, IP
-        RIGHTS ARE INTENDED TO ENCOURAGE INNOVATION AND THE
-        SPREAD OF KNOWLEDGE BY PROVIDING INCENTIVES TO
-        CREATE NEW WORKS AND GENERATE USEFUL INVENTIONS. IP
-        LAW MAY ALSO SERVE OTHER PURPOSES, SUCH AS PROMOTING
-        FAIR COMPETITION, PREVENTING CONSUMER CONFUSION, OR
-        DETERRING ECONOMIC ESPIONAGE.”
-      </p>
-      <p className='font-vista-sans py-2'>
-        <strong className='font-hwt-artz'>“ART”</strong> MEANS ANY DIGITAL CREATION, ART, DESIGN, AND
-        DRAWINGS CREATED BY ITS AUTHOR, AND FIRST RELEASED TO
-        THE PUBLIC BY THE CREATOR, THAT MAY BE ASSOCIATED
-        WITH AN NFT THAT CAN BE TRANSFERRED OR LICENSED TO AN
-        OWNER (SUBJECT TO THE LIMITATIONS SET FORTH HEREIN).
-      </p>
-      <p className='font-vista-sans py-2'>
-        <strong className='font-hwt-artz'>“ATTRIBUTE”</strong> MEANS EACH AND ALL CONSTITUENT
-        ELEMENTS OF THE ART, TAKEN INDEPENDENTLY OF THE
-        OVERALL ART, INCLUDING DESIGNS AND PATTERNS OF THE
-        BACKGROUND, APPAREL, EYES, AND SKIN TONES.
-      </p>
-      <p className='font-vista-sans py-2'>
-        <strong className='font-hwt-artz'>"NFT"</strong> MEANS ANY BLOCKCHAIN-TRACKED, NON-FUNGIBLE
-        TOKEN, SUCH AS THOSE CONFORMING TO THE ERC-721
-        STANDARD, ISSUED AND TRADABLE ON A BLOCKCHAIN.
-      </p>
-      <p className='font-vista-sans py-2'>
-        <strong className='font-hwt-artz'>“CREATOR”</strong> MEANS THE ENTITY WHICH RELEASED TO THE
-        PUBLIC THE PURCHASED NFT, AND RELATED ART, ON BEHALF
-        OF ITS AUTHOR, AND/OR ACQUIRED THOSE RIGHTS AND
-        INTERESTS, I.E., FROGGY FRIENDS LLC, WITH A REGISTERED
-        ADDRESS OF P.O. BOX 64101 IRVINE, CALIFORNIA USA.
-      </p>
-      <p className='font-vista-sans py-2'>
-        <strong className='font-hwt-artz'>“EFFECTIVE DATE”</strong> MEANS THE DATE ON WHICH THIS
-        AGREEMENT, AS DEFINED HEREINAFTER, SHALL BE DEEMED
-        ENTERING INTO FORCE, THAT IS (1) AS FOR THE PRIMARY
-        OWNER, THE DATE ON WHICH THIS AGREEMENT IS DISPLAYED
-        ON THE FROGGY FRIENDS WEBSITE INITIALLY AND THEREFORE
-        DEEMED ATTACHED TO THE PURCHASED NFT, AND (2) THE
-        DATE OF ACQUISITION OF THE PURCHASED NFT.
-      </p>
-    </div>
-  )
-}
